+++ NEWS.md
@@ -1,5 +1,11 @@
 # chart News
 
+## chart version 1.2.0
+
+- The functions to use viridis palettes in ggplot2 are cloned here from ggplot2
+  v. 3.0.0, in order to use them also with older ggplot2 plots.
+
+
 ## chart version 1.1.0
 
 - First implementation of base and lattice versions of `chart()`s.